Load http_archive and git_repository from Bazel tools
In newer versions of Bazel (>= 0.20 as best as I can tell) The native
`http_archive` and `git_repository` rules are disabled by default. Consequently,
this commit updates `repositories.bzl` to explicitly load the rules and updates
the calls to them accordingly. I also took this oppurtunity to clean up the
README.